requirements.txt
async-timeout==4.0.3
blinker==1.7.0
cachetools==2.0.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
google-api-python-client==1.7.2
google-auth==1.8.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.63.0
httplib2==0.22.0
idna==3.7
importlib_metadata==7.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
oauthlib==3.2.2
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
python-dotenv==1.0.1
redis==5.0.3
requests==2.31.0
requests-oauthlib==2.0.0
rsa==3.1.4
six==1.16.0
spotipy==2.23.0
uritemplate==3.0.0
urllib3==2.2.1
Werkzeug==3.0.2
zipp==3.19.2